+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: jmoreau <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/12/24 02:54:49 by jmoreau           #+#    #+#             */
-/*   Updated: 2015/02/06 00:40:23 by jmoreau          ###   ########.fr       */
+/*   Updated: 2015/02/12 04:45:09 by jmoreau          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t n)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (i++ < n)
